lines 0-1000
