+++ src/compiler/checker/mod.rs
@@ -1,5 +1,7 @@
-mod lines_0_2000;
-pub use lines_0_2000::*;
+mod lines_0_1000;
+pub use lines_0_1000::*;
+mod lines_1000_2000;
+pub use lines_1000_2000::*;
 mod lines_2000_4000;
 pub use lines_2000_4000::*;
 mod lines_4000_8000;
